Can't implement a catch-all in global namespace.
With a lot of hackery it is possible to make this work
even for enum's that already have a template function defined
for this that is found by ADL. But it doesn't work in all
cases as catch-all: if called from a namespace that already
has any operator<< defined, then enclosing namespaces are
not searched any further and we'd never reach global namespace
with unqualified lookup.

It could still work by putting this in namespace std, but I
feel that is too dirty for an important function like this
that in the end might be the way that anyone wants to write
many different enum types with.